Refactored unary operator, and provid support for logic NOT operator.
