+++ ODLA/platforms/dnnl/CMakeLists.txt
@@ -31,6 +31,7 @@ set(ODLA_DNNL_SRC
   odla_dnnl_cast.cc
   odla_dnnl_loss.cc
   odla_dnnl_rnn.cc
+  odla_dnnl_unary.cc
   odla_dnnl_binary.cc
   odla_dnnl_statistics.cc
 )
