+++ tests/unittests/lit_cases/test_dnnl/test_not_3d_dnnl.cc
@@ -25,5 +25,4 @@
 // RUN: %t_dnnl.exe 0.0001 0 dnnl %data_path/test_not_3d | FileCheck %s
 // CHECK: Result Pass
 // clang-format on
-// XFAIL: *
 #include "test_not_3d_dnnl.cc.tmp.main.cc.in"
